_libssh2_packet_requirev: get the len of the packet types vector as an argument
Even if the SSH protocol does not use packets of type 0 but _libssh2_packet_requirev
should not depend on that, it is the wrong layer, and anyway, a remote server may
send packets of type 0 and they must be handled correctly (ignored).

Until now the function was taking a null terminated string ("packet_types") as
the vector of acceptable packet types and using "strchr" to check if the type of the
received packet was included on it, but that would always match type 0 packets,
so an ugly explicit check for the type not being 0 was also in place.

This patch adds the length of the packet_types vector as an argument
to _libssh2_packet_requirev so that the special handling of type 0 could be
removed, and modifies the client code accordingly.

_libssh2_packet_askv has been modified accordingly, to accept the length of
the types vector as an argument too.

Null terminators has also been removed from the packet_types declarations in client
code, so that the "sizeof()" operator could be used to calculate the length
automatically.