_libssh2_packet_requirev: don't handle timeouts at this layer
This is the wrong layer to handle timeouts. Besides that, it was using
a hardcoded value (LIBSSH2_READ_TIMEOUT) instead of the user provided
one (session->api_timeout).

Removing that also means that _libssh2_packet_requirev is now a stateless
function, and so the state argument can be removed, and all the
packet_requirev_state_t slots scattered around the library structures
and the initialization code in the calling functions.