_libssh2_packet_askv, _libssh2_packet_ask: refactor
Move _libssh2_packet_ask logic into _libssh2_packet_askv and then
rewrite _libssh2_packet_ask to use _libssh2_packet_askv under the hood.

Also, allow matching any packet type by passing NULL as the
packet_types argument.

Note that this patch changes the semantics of libssh2_packet_askv slightly:
Before, the packet_types was used in the outer loop and now it is used in
the inner loop. That changes the order in which packets are found in the
brigade. The order implemented in this patch is the correct one.