libssh2_keepalive_send: fix EAGAIN handling
This function was ignoring LIBSSH2_ERROR_EAGAIN errors comming
from _libssh2_transport_send, but when that happens, the later
must be called again with exactly the same arguments.

This patch saves the keepalive packet data into the session structure
so that the transport send operation could be continued the next
time libssh2_keepalive_send is called.