_libssh2_wait_socket: error when block_directions is 0
When libssh2_session_block_directions returns zero, select/poll is
called with empty read/write sets and a timeout of 1 second. It
returns 0 and then _libssh2_wait_socket returns with a
LIBSSH2_ERROR_TIMEOUT. In other words, it waits for one second and
then returns with an unrelated error.

This patch, changes _libssh2_wait_socket to return
LIBSSH2_ERROR_BAD_USAGE instead.  It may uncover other hidden bugs on
the library that then should be fixed.