ssh2_echo: channel IO can not really be interleaved, fixes #167
When `libssh2_channel_read`|`write` returns `EAGAIN` the same function
must be called repeatedly with the same arguments until it returns
success or a different error but ssh2_echo was not honoring that
requirement.

This patch sets the session to blocking just before calling the
read/write methods in order to ensure it.