Fixed _libssh2_openssh_pem_parse_memory
Added a test to test to test libssh2_userauth_publickey_frommemory() using a new OpenSSL key type which showed the existing code wasn’t quite right (was expecting a parsed key, not a raw key). I also removed the header args as they are always the same with the new key format.