channel: Refactor closed tracking logic to be more robust.
Previously, there was a "close" flag in both the local and remote
state structures, but only the local one was checked in
_libssh2_channel_close, meaning that if someone invoked "channel_close"
after the channel had already been closed by the server,
an indefinite hang would occur.

Now, there is only one "closed" field, which is set when the
server replies that the channel is in fact closed.