ci: Checkout correct ref for ossfuzz
Up until now, this script has been setting the container to checkout the
target branch of a PR, usually master. This means that a PR is never
fuzzed before merging, which sort of defeats the purpose of CI on PRs.

Git clone does not allow random refs, only branches and tags, hence the
`--no-checkout` and the follow-up checkout.