sftp: Prevent files from being skipped if the output buffer is too small
LIBSSH2_ERROR_BUFFER_TOO_SMALL is returned if the buffer is too small
to contain a returned directory entry. On this condition we jump to the
label `end`. At this point the number of names left is decremented
despite no name being returned.

As suggested in #714, this commit moves the error label after the
decrement of `names_left`.

Fixes #714