add mbedTLS 3.x support
Make libssh2 compile cleanly with mbedTLS 3.x and later.

This patch makes use of `MBEDTLS_PRIVATE()`, which is not the
recommended, future-proof way to access mbedTLS data structures. This
method may break with a minor upgrade, according to the authors. This
is also the method used by libcurl.

Also:

- Fix a potentially uninitialized variable in
  `libssh2_mbedtls_rsa_sha2_sign()`. This happened in an error path,
  resulting in an unnecessary mbedTLS API call, with an uninitialized
  `md_type`.

- Bump mbedTLS version used in CI tests to 3.2.1.

Fixes #751