cmake: fix 'Policy CMP0080' warning
```
CMake Warning (dev) at C:/Program Files (x86)/CMake/share/cmake-3.16/Modules/BundleUtilities.cmake:233 (message):
  Policy CMP0080 is not set: BundleUtilities cannot be included at configure
  time.  Run "cmake --help-policy CMP0080" for policy details.  Use the
  cmake_policy command to set the policy and suppress this warning.

Call Stack (most recent call first):
  C:/Program Files (x86)/CMake/share/cmake-3.16/Modules/BundleUtilities.cmake:242 (_warn_cmp0080)
  tests/CMakeLists.txt:39 (include)
This warning is for project developers.  Use -Wno-dev to suppress it.
```

```
$ cmake --help-policy CMP0080
CMP0080
-------

.. versionadded:: 3.13

`BundleUtilities` cannot be included at configure time.

The macros provided by `BundleUtilities` are intended to be invoked
at install time rather than at configure time, because they depend on the
listed targets already existing at the time they are invoked. If they are
invoked at configure time, the targets haven't been built yet, and the
commands will fail.

This policy restricts the inclusion of `BundleUtilities` to
`cmake -P` style scripts and install rules. Specifically, it looks for the
presence of `CMAKE_GENERATOR` and throws a fatal error if it exists.

The `OLD` behavior of this policy is to allow `BundleUtilities` to
be included at configure time. The `NEW` behavior of this policy is to
disallow such inclusion.

This policy was introduced in CMake version 3.13.  CMake version
3.25.2 warns when the policy is not set and uses `OLD` behavior.
Use the `cmake_policy()` command to set it to `OLD` or `NEW`
explicitly.

.. note::
  The `OLD` behavior of a policy is
  `deprecated by definition`
  and may be removed in a future version of CMake.
```