always use SecureZeroMemory() on Windows
Even though the MSDN page of this function does not list it as UWP
compatible, this function is an alias for RtlSecureZeroMemory(). Doc [1]
for the latter says that the function is inlined, so it's safe to use
regardless of Windows version. mingw-w64-generated objects also confirm
that. libssh2 also had that unconditionally enabled in WinCNG with no
report of issues. It's likely dependent on compiler version, but more
half-decently modern ones should support it.

Make a leap and use it unconditionally in _libssh2_explicit_zero().

[1] https://learn.microsoft.com/en-us/windows-hardware/drivers/ddi/wdm/nf-wdm-rtlsecurezeromemory