agent: fix compiler warning in snprintf() call
from gcc:
```
~/libssh2/src/agent.c:296:35: warning: 'snprintf' output truncated before the last format character [-Wformat-truncation=]
  296 |              "PageantRequest%08x%c", (unsigned)GetCurrentThreadId(), '\0');
      |                                   ^
```

Remove this tricky code, that aimed to ensure a terminating zero by
adding an extra one, for Microsoft platforms where we map `snprintf()`
to `_snprintf()`. The latter is famous for not guaranteeing
a terminating zero.

However, this was the single instance in libssh2 with this trick, even
though this particular case doesn't seem to be affected, as the output
shall fit into the buffer always, one terminating zero included.

Refs:

- https://learn.microsoft.com/en-us/cpp/c-runtime-library/reference/snprintf-snprintf-snprintf-l-snwprintf-snwprintf-l

  "The _snprintf family of functions only appends a terminating null
  character if the formatted string length is strictly less than count
  characters."

- https://learn.microsoft.com/en-us/previous-versions/visualstudio/visual-studio-2012/2ts7cx93(v=vs.110)