cmake: unset forced CMAKE_C_STANDARD 90 (#822)
Added in cf80f2f4b5255cc85a04ee43b27a29c678c1edb1 (on 2016-08-14),
with the title "Basic dockerised test suite".

It's not clear why a C standard was explicitly set, but a side-effect
of this is that CMake-built binaries diverged from ones built with
autotools or GNU Make (using the same compiler and configuration).

Another issue is that this may introduce ABI incompatibility with
binaries built with a different C standard flag, e.g. the C compiler
default or one used for other components of a final app.

Seems unlikely, but if our tests require this option, we should set it
for the CI builds only?