cmake: extend workaround for linking a test with shared libssh2
One of the tests uses internal libssh2 functions, and with CMake it
compiles `misc.c` directly to solve it. `misc.c` references bcrypt /
blowfish code. This needs a workaround for build configs where libssh2
doesn't export these.

Before this patch, we enabled this workaround for MSVC.

In the patch we extend this to all Windows. There is no CI test for
this, but gcc and llvm/clang + mingw64 builds also need it. Most likely
all Windows, and possibly other systems, so also make it easy to enable
this at a single point.

[ autotools builds force-link this one test against static libssh2. ]