example: replace strcasecmp() with C89 strcmp()
Another option is to duplicate the check_symbol_exists() workaround
from src/CMakeLists.txt, but I figure it's not worth the complexity.
We use this function solely to check optional command-line options
for example programs, and those are just fine as lower-case.

Otherwise MSVC CI jobs fail with ENABLE_WERROR.