build: add autotools support for `libssh2.rc`
Caveat: When building `--enable-static` and `--enable-shared` at the
same time, the compiled Windows resource is also included in the
static library. This appears to be an autotools limitation, where there
it seems to offer no way to have a different input list (or different
custom options) for shared and static libraries, even though autotools
builds them separately.

The workaround is to build static libraries in a separate
`./configure` + `make` pass.

Closes #xxx