transport: rename local `RANDOM_PADDING` macro
Rename `RANDOM_PADDING` macro used internally to enable some code.

Committed in the initial version of `transport.c` in
9d55db6501aa4e21f0858cf36cdc2ddc11b96e83 (2007-02-02). libssh2 code
never defined it.

The name happens to collide with a Windows macro in `wincrypt.h`.
`transport.c` doesn't include this header, but it includes `winsock2.h`,
and it turns out it can also define this macro in some cases, e.g.
when `WIN32_LEAN_AND_MEAN` is not set.

To be on the safe side, prefix the name with `LIBSSH2_` to avoid
enabling it by accident.

Q: Maybe it'd be best to delete it with the guarded code?

Reported-by: Markus-Schmidt on github
Fixes #921
Closes #924