build: support `libssh2.rc` with autotools
Caveat: When building `--enable-static` and `--enable-shared` at the
same time, the compiled Windows resource is also included in the
static library. This appears to be an autotools limitation, with no
way to have different input lists (or different custom options) for
shared and static libraries, even though it builds them separately.

The workaround is to build static libraries in a separate
`./configure` + `make` pass.

Closes #944