delete `--quiet` option from command `docker pull`
This option requires docker v19.03, and AppVeyor's Visual Studio 2017
image doesn't support it.

Log output did not change without `--quiet`, so it seems safe to delete
it.

In case we'd need it, another solution is to retry without `--quiet`
if the command fails. docker's exit status is 125 in that case.