tidy-up: drop no-op use of `getcwd` and `sys/param.h`
If not `FIXTURE_WORKDIR` envvar of compile-time macro is present,
just don't do anything. Before this patch, we retrieved the current
directory and changed directory into it, which is pointless.

This also makes `sys/param.h` unnecessary to detect and use, and
we can drop an MSVC-specific hack.

Delete an extra idle use of the above header.

Still remains the two methods of setting `FIXTURE_WORKDIR`, of which
only the compile-time method is actually used and only with CMake,
for an Xcode workaround.

Ref: https://github.com/libssh2/libssh2/pull/198/commits/10a5cbf945abcc60153ee3d59284d09fc64ea152
Ref: 9f8e69cca0b94266c174ae13a860a02c762a3f4a

Closes #xxxx