vms: fix to include `sys/socket.h` [ci skip]
Due to a typo in the `HAVE_*` macro, this header was never included.

A comment suggests that `socklen_t` is not defined on VMS and defines it
manually. This symbol is usually in `sys/socket.h`, so the typo may well
have been the reason for it to be missing.

Closes #1007