test_sshd.test: integrate sshd init/start/stop
And stop relying on `sshd_fixture.sh` for this.

It means that we now do a single sshd init/start for the whole list
of tests and shutdown the server afterwards. Instead of start/stopping
it for each tests. The latter turned out to not work reliably or at all,
on Windows. An sshd server should handle our test request in a single
server session, even in parallel.

Also make this script compatible with `sshd_fixture.sh` when passing
a single executable to it. Extend this to support passing any number of
tests via the command-line.

TODO: unlock all tests on Windows.
TODO: try doing the same under CMake, which doesn't use the
      `test_sshd.test` script yet.