autotools: delete explicit `builddir`
Added in Added in 13f8addd1bc17e6c55d52491cc6304319ac38c6d (2015-07-02)

It seems this wasn't necessary to make the above fix work, and
`mansyntax.sh` is able to figure out the build workdir by reading
$PWD. Our out-of-tree and `make distcheck` CI builds also work
without it.

Let us know if there is a scenario we're missing and needs this.