cmake: re-add `select()` detection (regression)
`select()` detection suffered two regressions: First I accidentally
deleted it for non-Windows [1]. Then the Windows-specific setting got
missed from the generated `libssh2_config.h` after a `CMakeLists.txt`
modification.

[1] 31fb8860dbaae3e0b7d38f2a647ee527b4b2a95f (2023-03-07)
[2] 803f19f004eb6a5b525c48fff6f46a493d25775c (2023-04-18)

This patch restores detection. For Windows it enables unconditionally,
not only for speed reasons, but because detection needs `ws2_32`, and
even that is broken on the x86 platform due to a calling convention
mismatch (as noted in the original `cmake/SocketLibraries.cmake`).

Closes #1031