cmake: allow tests with `BUILD_STATIC_LIBS=OFF`
Before this patch, the CMake build did not allow to disable static
libssh2 library while also building tests.

This patch removes this constraint, and makes this combination possible.
In this case the 3 (at the moment) tests that require a static libssh2
library, are skipped from the build and test runs.

Cherry-picked from #1036